2D -Vector
This code presents a solution to a HackerRank problem where a 2D vector is declared. A nested loop is employed to take user inputs for each row of the vector, and the elements are stored in a temporary vector. The temporary vector is then inserted at the Kth position in the 2D vector. Finally, a loop retrieves and prints elements from specific positions in the 2D vector based on user queries.